+++ server/router/index.js
@@ -20,31 +20,35 @@
 // THE SOFTWARE.
 
 const Router = require('koa-router');
-const losslessJSON = require('lossless-json');
-const moment = require('moment');
 
-const featureFlags = require('./feature-flags.json');
 const {
-  buildQueryString,
+  domainAuthorizationHandler,
+  domainHandler,
+  domainListHandler,
+  featureFlagHandler,
+  healthHandler,
+  tasklistHandler,
+  tasklistPartitionListHandler,
+  tasklistPollerListHandler,
+  workflowArchivedListHandler,
+  workflowExportHandler,
+  workflowHandler,
+  workflowHistoryHandler,
+  workflowListHandler,
+  workflowQueryHandler,
+  workflowQueryTypeHandler,
+  workflowSignalHandler,
+  workflowTerminateHandler,
+} = require('./routes');
+
+const {
   listWorkflows,
-  mapHistoryResponse,
 } = require('./helpers');
 
 const router = new Router();
 
-router.get('/api/domains', async function(ctx) {
-  ctx.body = await ctx.cadence.listDomains({
-    pageSize: 50,
-    nextPageToken: ctx.query.nextPageToken
-      ? Buffer.from(ctx.query.nextPageToken, 'base64')
-      : undefined,
-  });
-});
-
-router.get('/api/domains/:domain', async function(ctx) {
-  ctx.body = await ctx.cadence.describeDomain({ name: ctx.params.domain });
-});
-
+router.get('/api/domains', domainListHandler);
+router.get('/api/domains/:domain', domainHandler);
 
 /**
  * Override this route to perform authorization check
@@ -61,13 +65,7 @@ router.get('/api/domains/:domain', async function(ctx) {
  *  };
  * })
  */
-router.get('/api/domains/:domain/authorization', async function(ctx, next) {
-  ctx.body = {
-    authorization: true,
-  };
-
-  next();
-});
+router.get('/api/domains/:domain/authorization', domainAuthorizationHandler);
 
 router.get(
   '/api/domains/:domain/workflows/open',
@@ -79,282 +77,53 @@ router.get(
   listWorkflows.bind(null, 'closed')
 );
 
-router.get('/api/domains/:domain/workflows/archived', async function(ctx) {
-  const { nextPageToken, ...query } = ctx.query || {};
-  let queryString;
+router.get('/api/domains/:domain/workflows/archived', workflowArchivedListHandler);
 
-  if (query.queryString) {
-    queryString = query.queryString;
-  } else {
-    const startTime = moment(query.startTime || NaN);
-    const endTime = moment(query.endTime || NaN);
-
-    ctx.assert(startTime.isValid() && endTime.isValid(), 400);
-    queryString = buildQueryString(startTime, endTime, query);
-  }
-
-  ctx.body = await ctx.cadence.archivedWorkflows({
-    query: queryString,
-    nextPageToken: nextPageToken
-      ? Buffer.from(nextPageToken, 'base64')
-      : undefined,
-  });
-});
-
-router.get('/api/domains/:domain/workflows/list', async function(ctx) {
-  const q = ctx.query || {};
-
-  ctx.body = await ctx.cadence.listWorkflows({
-    query: q.queryString || undefined,
-    nextPageToken: q.nextPageToken
-      ? Buffer.from(q.nextPageToken, 'base64')
-      : undefined,
-  });
-});
+router.get('/api/domains/:domain/workflows/list', workflowListHandler);
 
 router.get(
   '/api/domains/:domain/workflows/:workflowId/:runId/history',
-  async function(ctx) {
-    const q = ctx.query || {};
-
-    ctx.body = await ctx.cadence.getHistory({
-      nextPageToken: q.nextPageToken
-        ? Buffer.from(q.nextPageToken, 'base64')
-        : undefined,
-      waitForNewEvent: 'waitForNewEvent' in q ? true : undefined,
-    });
-
-    ctx.body.history.events = mapHistoryResponse(ctx.body.history);
-  }
+  workflowHistoryHandler
 );
 
 router.get(
   '/api/domains/:domain/workflows/:workflowId/:runId/export',
-  async function(ctx) {
-    let nextPageToken;
-
-    do {
-      const page = await ctx.cadence.exportHistory({ nextPageToken });
-
-      if (!nextPageToken) {
-        ctx.status = 200;
-      }
-
-      ctx.res.write(
-        (nextPageToken ? ',' : '[') +
-          page.history.events.map(losslessJSON.stringify).join(',')
-      );
-      nextPageToken =
-        page.nextPageToken && Buffer.from(page.nextPageToken, 'base64');
-    } while (nextPageToken);
-
-    ctx.res.write(']');
-    ctx.body = '';
-  }
+  workflowExportHandler
 );
 
 router.get(
   '/api/domains/:domain/workflows/:workflowId/:runId/query',
-  async function(ctx) {
-    // workaround implementation until https://github.com/uber/cadence/issues/382 is resolved
-    try {
-      await ctx.cadence.queryWorkflow({
-        query: {
-          queryType: '__cadence_web_list',
-        },
-      });
-
-      ctx.throw(500);
-    } catch (e) {
-      ctx.body = ((e.message || '').match(
-        /(KnownQueryTypes|knownTypes)=\[(.*)\]/
-      ) || [null, null, ''])[2]
-        .split(/, | /)
-        .filter(q => q);
-    }
-  }
+  workflowQueryHandler
 );
 
 router.post(
   '/api/domains/:domain/workflows/:workflowId/:runId/query/:queryType',
-  async function(ctx) {
-    ctx.body = await ctx.cadence.queryWorkflow({
-      query: {
-        queryType: ctx.params.queryType,
-      },
-    });
-  }
+  workflowQueryTypeHandler
 );
 
 router.post(
   '/api/domains/:domain/workflows/:workflowId/:runId/terminate',
-  async function(ctx) {
-    ctx.body = await ctx.cadence.terminateWorkflow({
-      reason: ctx.request.body && ctx.request.body.reason,
-    });
-  }
+  workflowTerminateHandler
 );
 
 router.post(
   '/api/domains/:domain/workflows/:workflowId/:runId/signal/:signal',
-  async function(ctx) {
-    ctx.body = await ctx.cadence.signalWorkflow({
-      signalName: ctx.params.signal,
-    });
-  }
+  workflowSignalHandler
 );
 
-router.get('/api/domains/:domain/workflows/:workflowId/:runId', async function(
-  ctx
-) {
-  try {
-    const describeResponse = await ctx.cadence.describeWorkflow();
+router.get('/api/domains/:domain/workflows/:workflowId/:runId', workflowHandler);
 
-    if (describeResponse.workflowExecutionInfo) {
-      describeResponse.workflowExecutionInfo.closeEvent = null;
-
-      if (describeResponse.workflowExecutionInfo.closeStatus) {
-        const closeEventResponse = await ctx.cadence.getHistory({
-          HistoryEventFilterType: 'CLOSE_EVENT',
-        });
-
-        describeResponse.workflowExecutionInfo.closeEvent = mapHistoryResponse(
-          closeEventResponse.history
-        )[0];
-      }
-    }
-
-    ctx.body = describeResponse;
-  } catch (error) {
-    if (error.name !== 'NotFoundError') {
-      throw error;
-    }
-
-    const archivedHistoryResponse = await ctx.cadence.getHistory();
-    const archivedHistoryEvents = mapHistoryResponse(
-      archivedHistoryResponse.history
-    );
-
-    if (!archivedHistoryEvents.length) {
-      throw error;
-    }
-
-    const { runId, workflowId } = ctx.params;
-
-    const {
-      timestamp: startTime,
-      details: {
-        taskList,
-        executionStartToCloseTimeoutSeconds,
-        taskStartToCloseTimeoutSeconds,
-        workflowType: type,
-      },
-    } = archivedHistoryEvents[0];
-
-    ctx.body = {
-      executionConfiguration: {
-        taskList,
-        executionStartToCloseTimeoutSeconds,
-        taskStartToCloseTimeoutSeconds,
-      },
-      workflowExecutionInfo: {
-        execution: {
-          runId,
-          workflowId,
-        },
-        isArchived: true,
-        startTime,
-        type,
-      },
-      pendingActivities: null,
-      pendingChildren: null,
-    };
-  }
-});
-
-router.get('/api/domains/:domain/task-lists/:taskList/pollers', async function(
-  ctx
-) {
-  const descTaskList = async taskListType =>
-    (
-      await ctx.cadence.describeTaskList({
-        domain: ctx.params.domain,
-        taskList: { name: ctx.params.taskList },
-        taskListType,
-      })
-    ).pollers || [];
-
-  const r = type => (o, poller) => {
-    const i = o[poller.identity] || {};
-
-    o[poller.identity] = {
-      lastAccessTime:
-        !i.lastAccessTime || i.lastAccessTime < poller.lastAccessTime
-          ? poller.lastAccessTime
-          : i.lastAccessTime,
-      taskListTypes: i.taskListTypes ? i.taskListTypes.concat([type]) : [type],
-    };
-
-    return o;
-  };
-
-  const activityL = await descTaskList('Activity'),
-    decisionL = await descTaskList('Decision');
-
-  ctx.body = activityL.reduce(
-    r('activity'),
-    decisionL.reduce(r('decision'), {})
-  );
-});
+router.get('/api/domains/:domain/task-lists/:taskList/pollers', tasklistPollerListHandler);
 
 router.get(
   '/api/domains/:domain/task-lists/:taskList/partitions',
-  async function(ctx) {
-    const { domain, taskList } = ctx.params;
-
-    ctx.body = await ctx.cadence.listTaskListPartitions({
-      domain,
-      taskList: { name: taskList },
-    });
-  }
+  tasklistPartitionListHandler
 );
 
-router.get('/api/feature-flags/:key', (ctx, next) => {
-  const {
-    params: { key },
-  } = ctx;
-  const featureFlag = featureFlags.find(featureFlag => featureFlag.key === key);
-  const value = (featureFlag && featureFlag.value) || false;
+router.get('/api/feature-flags/:key', featureFlagHandler);
 
-  ctx.body = {
-    key,
-    value,
-  };
+router.get('/api/domains/:domain/task-lists/:taskListName', tasklistHandler);
 
-  next();
-});
-
-router.get('/api/domains/:domain/task-lists/:taskListName', async function(
-  ctx
-) {
-  const { domain, taskListName } = ctx.params;
-  const descTaskList = async taskListType =>
-    await ctx.cadence.describeTaskList({
-      domain,
-      taskList: { name: taskListName },
-      taskListType,
-    });
-
-  const activityList = await descTaskList('Activity');
-  const decisionList = await descTaskList('Decision');
-  const activityPollerList = activityList.pollers || [];
-  const decisionPollerList = decisionList.pollers || [];
-
-  const taskList = { pollers: [...activityPollerList, ...decisionPollerList] };
-
-  ctx.body = taskList;
-});
-
-router.get('/health', ctx => (ctx.body = 'OK'));
+router.get('/health', healthHandler);
 
 module.exports = router;
