+++ server/router/routes/domain-authorization-handler.js
@@ -1,0 +1,9 @@
+const domainAuthorizationHandler = async (ctx, next) => {
+  ctx.body = {
+    authorization: true,
+  };
+
+  next();
+};
+
+module.exports = domainAuthorizationHandler;
