+++ server/router/routes/domain-handler.js
@@ -1,0 +1,5 @@
+const domainHandler = async (ctx) => {
+  ctx.body = await ctx.cadence.describeDomain({ name: ctx.params.domain });
+};
+
+module.exports = domainHandler;
