+++ server/router/routes/domain-list-handler.js
@@ -1,1 +1,10 @@
-// TODO
+const domainListHandler = async (ctx) => {
+  ctx.body = await ctx.cadence.listDomains({
+    pageSize: 50,
+    nextPageToken: ctx.query.nextPageToken
+      ? Buffer.from(ctx.query.nextPageToken, 'base64')
+      : undefined,
+  });
+};
+
+module.exports = domainListHandler;
