+++ server/router/routes/health-handler.js
@@ -1,0 +1,3 @@
+const healthHandler = ctx => (ctx.body = 'OK');
+
+module.exports = healthHandler;
