+++ server/router/routes/tasklist-partition-list-handler.js
@@ -1,0 +1,10 @@
+const tasklistPartitionListHandler = async (ctx) => {
+  const { domain, taskList } = ctx.params;
+
+  ctx.body = await ctx.cadence.listTaskListPartitions({
+    domain,
+    taskList: { name: taskList },
+  });
+};
+
+module.exports = tasklistPartitionListHandler;
