+++ server/router/routes/workflow-query-type-handler.js
@@ -1,0 +1,9 @@
+const workflowQueryTypeHandler = async (ctx) => {
+  ctx.body = await ctx.cadence.queryWorkflow({
+    query: {
+      queryType: ctx.params.queryType,
+    },
+  });
+};
+
+module.exports = workflowQueryTypeHandler;
