+++ server/router/routes/workflow-signal-handler.js
@@ -1,0 +1,7 @@
+const workflowSignalHandler = async (ctx) => {
+  ctx.body = await ctx.cadence.signalWorkflow({
+    signalName: ctx.params.signal,
+  });
+};
+
+module.exports = workflowSignalHandler;
