+++ server/router/routes/workflow-terminate-handler.js
@@ -1,0 +1,7 @@
+const workflowTerminateHandler = async (ctx) => {
+  ctx.body = await ctx.cadence.terminateWorkflow({
+    reason: ctx.request.body && ctx.request.body.reason,
+  });
+};
+
+module.exports = workflowTerminateHandler;
